claim-explanation-bot/requirements.txt
streamlit==1.32.2
python-dotenv==1.0.1
pdfplumber==0.10.4

langchain==0.1.16
langchain-openai==0.0.8
